backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
PyMySQL==1.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
marshmallow==3.20.1
scikit-learn==1.3.0
pandas==2.1.1
numpy==1.24.3
APScheduler==3.10.4
python-dateutil==2.8.2
email-validator==2.0.0
Pillow==10.0.1
celery==5.3.2
redis==4.6.0
gunicorn==21.2.0
